refactor(editor): Extract ToolCallStep component from AgentTimeline
Extract the duplicated tool call step rendering (icon column, connector,
label, collapsible data toggle) from AgentTimeline.vue and
SubagentStepTimeline.vue into a shared ToolCallStep component. This
removes ~100 lines of duplicated template and CSS.
